+++ src/Stack/Options/BuildMonoidParser.hs
@@ -18,8 +18,9 @@ buildOptsMonoidParser hide0 =
     libProfiling <*> exeProfiling <*> libStripping <*>
     exeStripping <*> haddock <*> haddockOptsParser hideBool <*>
     openHaddocks <*> haddockDeps <*> haddockInternal <*>
-    haddockHyperlinkSource <*> copyBins <*> preFetch <*> keepGoing <*>
-    forceDirty <*> tests <*> testOptsParser hideBool <*> benches <*>
+    haddockHyperlinkSource <*> copyBins <*> copyCompilerTool <*>
+    preFetch <*> keepGoing <*> forceDirty <*>
+    tests <*> testOptsParser hideBool <*> benches <*>
     benchOptsParser hideBool <*> reconfigure <*> cabalVerbose <*> splitObjs <*> skipComponents
   where
     hideBool = hide0 /= BuildCmdGlobalOpts
@@ -62,7 +63,6 @@ buildOptsMonoidParser hide0 =
                       \debuggers/profiling tools/other utilities that use \
                       \debugging symbols." <>
              hideExceptGhci)
-
     libProfiling =
         firstBoolFlags
             "library-profiling"
@@ -109,6 +109,11 @@ buildOptsMonoidParser hide0 =
         firstBoolFlags
             "copy-bins"
             "copying binaries to the local-bin-path (see 'stack path')"
+            hide
+    copyCompilerTool =
+        firstBoolFlags
+            "copy-compiler-tool"
+            "copying binaries of targets to compiler-tools-bin (see 'stack path')"
             hide
     keepGoing =
         firstBoolFlags